changed constants for table size

+++ src/edu/upenn/cis/cis350/database/RecentSearches.java
@@ -195,12 +195,12 @@ public class RecentSearches extends DatabaseHelperClass {
 	 * @param type - 0 for recent, 1 for favorites
 	 * @return
 	 */
-	public long getSize(int type) {
+	public double getSize(int type) {
 		String table = (type == 0) ? SEARCHES_TABLE : FAVORITES_TABLE;
 		Cursor c = mDb.rawQuery("SELECT COUNT(*) as num FROM " + table , null);
 		c.moveToFirst();
 		int num = c.getInt(c.getColumnIndex("num"));
 		
-		return num * Constants.FAVORITE_ROW_SIZE / 1000;
+		return ((double)num * Constants.FAVORITE_ROW_SIZE) / 1000;
 	}
 }
